Fixes DCHECK crash in SharedPinnedTabServiceBrowserTest.ActivatePinnedTab
The crash was happening because on closing the last browser we were
adding the cached shared pinned contents to it via
AddRestoredTabFromCache with the select param set to true. Having select
set to true lead to activating the tab and
SharedPinnedTabService::OnBrowserSetLastActive getting called, which
DCHECKs because the activating browser has already been removed from
browsers_ list.

The solution here is to not set select to true when restoring the pinned
tab on the last browser closing.