Try fixing crash in shortcut parsing
According to callstack, the following line seems to be the cause of the crash:

```
  const unsigned short charCode =
      [[keyEquivalent uppercaseString] characterAtIndex:0];
```

As we're already checking keyEquivalent's length, try checking
length of uppercaseString.

Also we're changing localization option for NSMenu and that could have
affected it. So try checking if we can get shortcut AFTER changing
the option.