[Android] Changes for logo polish for tablet
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/4a6ecfb34aa32438b180284a909dcd2db639ce8e

[LogoPolish] Implement logo polish for tablet.

This CL modifies the logo height on tablet when Logo Polish is
activated. The height of the logo is decided by the selected variation.
When using split screen mode on a tablet, the logo's height will remain
at the small setting.

The screenshots are included in the following link:
https://docs.google.com/document/d/16z7y1lXiSJ0sCZ0-UzrXdBC8H2idDrVdVarJ4DIEuho/edit?usp=sharing

Bug: 328769923, 330745894