[Android] Changes for ExternalNavigationHandler startActivity
Chromium change:
https://github.com/brave/chromium/commit/e5167d720ffaaae5565500d1bc62593eb3e266c5

Do not allow repeatedly launching apps for a single user activation
If a site repeatedly navigates through javascript, the first navigation
consumes the user gesture, but subsequent navigations are considered as
on the same navigation chain and so don't need a gesture. We should
make sure to prevent these subsequent navigations from launching apps
if either a previous app was already launched, or failed to launch due
to the app not being installed (so this can't be used as a
fingerprinting vector).

Bug: 330555390