Rebroadcast Solana transactions every 5 seconds on client side
This is done to increase the successful rate of Solana transactions, we
rebroadcast on client side rather than sorely depend on remote RPC node's retry
queue. This is safe because the same exact transaction (same signature) will
not be accepted by the network twice.
Also add a 30 minutes threshold to drop old transactions.