Disables DigitalGoodsAPIBrowserTest browser test.
We have digital goods API disabled in
BraveContentRendererClient::SetRuntimeFeaturesDefaultsBeforeBlinkInitialization.
With /0 variation the test works as expected and the API is not
available.

However with /1 variation the test runs with the
--enable-blink-test-features switch and the API should become available,
but it doesn't.

Running brave Nightly with cr124 and the switch enables the API as can
be tested by opening devtools and running window.getDigitalGoodsService().
The same doesn't work with cr125.

Created a followup issue to investigate.