`WebstoreInstallWithPrompt` deleted one ctor
A constructor has been deleted in this class, leaving the type to be
instantiated only with another constructor that requires a
`parent_window` to be provided. As the parent window can be passed as
`nullptr`, this change uses this constructor with `nullptr` for the
window, to preserve all the existing behaviour.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/0bf35b00aade6855fc8df8e2f6b1c594da02d45a

commit 0bf35b00aade6855fc8df8e2f6b1c594da02d45a
Author: Devlin Cronin <rdevlin.cronin@chromium.org>
Date:   Fri May 10 20:47:18 2024 +0000

    [Extensions] Remove unused ctor on WebstoreInstallWithPrompt

    WebstoreInstallWithPrompt had an unused ctor. Remove it.

    Bug: None