[CodeHealth] Use `span` for `SecureZeroData`
`SecureZeroData` is used to clean up memory in a way the data is zeroed
out once out of scope. The use of a custom allocator helps to enforce
that. However the current way buffers are handled runs into issues when
using `Wunsafe-buffers-usage`.

This change changes the interface of this function to use `span` to
communicate the buffer. Additionally, this change simplifies the way
different platforms are handled by having a single handling for all of
them.

Finally, this change corrects the UB in the tests, where instatiated
types like `vector`, and `string` were being menset to naught.