[iOS] Call WebKit handlers when alerts are dismissed without user input
This introduces changes to the alert controllers used for both JavaScript alerts and to media permission prompts so that in the case where a user may inadvertently dismiss the prompts without selecting an option (for instance by backgrounding the app while they're visible and then opening a new link from outside of Brave, or activating a Brave widget shortcut) is properly handled and the necessary WebKit completion handlers are called.

This also reverts an attempted fix (#23701), as this introduced a new way of not calling completion handlers.