[Android] Use BottomControlsStacker to resize bottom browser controls
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/d2c4b3d248d0a74dd00e9956df65604eea0ef6bc

[E2E] Use BottomControlsStacker to resize bottom browser controls

Introduce BrowserControlsStacker to coordinate different UI that wants to change the bottom controls height. This CL is used as transitional solution, just forwarding calls from BrowserControlsStacker to the wrapped BrowserControlsSizer, while also provide the BrowserControlsStateProvider for clients to do the observation.

There are still references of BrowserControlsSizer in custom_tabs, which  is not in scope of this initial refactor.

Bug: 345488108