Add allow-same-origin attribute to sandbox attribute for portfolio chart
Since the IsolateSandboxedIframes feature flag was enabled, the
chrome-untrusted://line-chart-display Wallet browser tests for the portfolio
chart have been failing. In order to temporarily work around those failures,
we've added allow-same-origin to the sandbox attribute for that chart.

Entered https://github.com/brave/brave-browser/issues/38771 to address this in a
follow-up.

Chromium change:
https://source.chromium.org/chromium/chromium/src/+/b2c49e9ef800bc606041fc9356e665d11869cb30

commit b2c49e9ef800bc606041fc9356e665d11869cb30
Author: W. James MacLean <wjmaclean@chromium.org>
Date:   Wed May 15 19:54:12 2024 +0000

    Enable IsolateSandboxedIframes by default.

    This CL enables the IsolateSandboxedIframes feature by default.

    Bug: 40082497