Remove draggable space above tabs to prevent accidental drag gestures
The draggable space above tabs constantly leads to accidental drag gestures and misplaced windows. 

This issue is particularly prevalent on smaller screens or when using touchpads, where the clickable area is already very limited. 

This 'feature' was never present prior to the last few weeks (in nightly builds).

It's simply not needed given due to the fact that even before the refresh users had (and still have) several other ways to drag windows if they wanted to.

By removing this draggable space, we can reduce the likelihood of users inadvertently dragging windows. Thank you.