[Android] Adjustments for the new tab switcher UI
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/c717fb07aed0f4756b9306ae0e55fa8f897bb1dc

[Hub] Cleanup legacy tab switcher code in toolbar

The feature flag controlling Hub was removed and isHubEnabled defaults
to true. Tests were already updated see:
https://chromium-review.googlesource.com/c/chromium/src/+/5634047

Toolbar no longer has a tab switcher mode as the toolbar is now
overlaid when the Hub is visible. Remove the legacy toolbar code related
to this now deprecated UI.

This is a follow up to
https://chromium-review.googlesource.com/c/chromium/src/+/5666417
which removed the StartSurface related code and significantly simplified
this cleanup.

Fixed: 350704312