[CodeHealth] Remove uses of `base::SupportsWeakPtr`
In upcoming chromium version, `base::SupportsWeakPtr` will be marked as
deprecated, and eventually removed. This class is being removed as it
leaves the weak ptr factory to be invalidated last, which can cause the
us of half-destroyed objects.

This change removes the use of the deprecated base class in favour of
composition, by having the factory as a member, and marking the class as
`final`, when possible, to make sure that the factory is the invalidated
first during destruction.

Some classes in this PR were deriving from `SupportsWeakPtr`, and yet
making no use of weak ptrs, and in these cases, all weak ptr references
are being removed.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/fe3114c4af662797ee08f1580951c4037e507ac1

commit fe3114c4af662797ee08f1580951c4037e507ac1
Author: David Bertoni <dbertoni@chromium.org>
Date:   Thu Jul 11 17:21:14 2024 +0000

    [Code Health] Add deprecation comment for base::SupportsWeakPtr.

    This class is deprecated, per the Code Health bug, and multiple
    reviewers have requested a comment to help avoid its use in new
    code.

    In addition, the constructor of this class is now private and the
    remaining uses are allowed via friend declarations. These will be
    cleaned up as the remaining uses are fixed.

    Bug: 40485134