Move Cast... menu item back under Save and share submenu.
Previously, we moved this item to under "More tools", but now want to
keep it where upstream has it, under "Cast, save, and share", except we
want to always keep the name "Save and share" regardless of whether cast
is shown in there or not. If the Media Router setting under
brave://settings/extensions is turned off the "Cast" menu item is not
shown.