Fixing gn `deps` dependency failures for buildflags
It looks like some of these dependencies were working incidently, due to
the build order of the targets, and they stopped working suddenly as the
build order may have changed for unknown reasons, which has exposed the
failure of certain dependencies to be specified.

This change corrects these missing dep inclusions.