Remove the allow-same-origin bypass needed for sandbox isolated frames
Essentially, because we didn't allow the wallet and wallet panel to
be able to request the chrome-untrusted:// frame, when the sandbox
isolated frames feature was turned on it made it so that the two
pages could no longer communicate since they were separated. Once
this enabled requesting it appeared to fix the tests, so that we
no longer need to add allow-same-origin to bypass the bug.