[canary] Print json output for failed patches
This change adds an argument a command line argument to `apply_patches`,
namely `--print-patch-failures-in-json`, that allows the output of a
JSON list of files that failed to apply. This will make it easier for us
to read the output from this script when running `scritp/version_up.py`,
and with this information we can further improve the automation of the
daily canary build.

A sample output:

[
  {
    patchPath: 'patches/chrome-android-java-src-org-chromium-chrome-browser-sync-settings-ManageSyncSettings.java.patch',
    path: 'chrome/android/java/src/org/chromium/chrome/browser/sync/settings/ManageSyncSettings.java'
  },
  {
    patchPath: 'patches/chrome-browser-ui-views-toolbar-toolbar_view.cc.patch',
    path: 'chrome/browser/ui/views/toolbar/toolbar_view.cc'
  }
]