Avoid `vector` copies, and outptrs in `FuzzyFinder`
In this code there were a few `vector<int>` container copies that should
be avoided, and have been replaced with `span`.

Additionally, the convention of passing `std::vector<gfx::Range>` as an
outptr should be avoided, in favour of a reference, as per chromium
guidelines, as a pointer opens the door for `nullptr` to be passed by
accident.