[Android] Adjustment to extending the omnibox suggestions when drawing E2E
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/3c34eed14f29d850c9756a3f7e1ef4f7cb4da709

Extend the omnibox suggestions when drawing E2E

By default, the omnibox suggestions fit the height of the screen, without including system bars. When the edge-to-edge feature is enabled, the bottom navigation bar region can be drawn into, showing a gap with tab content underneath the omnibox suggestions list. This change allows for the edge-to-edge feature to communicate extra bottom padding to the omnibox suggestions dropdown to allow it to extend down and cover that bottom region, so the dropdown can draw "edge-to-edge" while keeping all of the suggestions fully visible and interactable.

Demo:
https://screencast.googleplex.com/cast/NjYyMDM3MTIxNzI4NTEyMHwzOTA4ZmJkZi00NQ

Demo (landscape):
https://screencast.googleplex.com/cast/NTk5NTYyMjA4NjczNzkyMHw5MzI3Y2ZlYi02YQ

Drawing E2E:
https://screenshot.googleplex.com/BUEWNAUMbguNCNY.png

Drawing E2E (bottom):
https://screenshot.googleplex.com/BuCyb2EzK95bsuC.png

Bug: b/325636004