[ubsan] Fix number overflow inside base10 funcs
This PR fixes `int256_t`/`uint256_t` overflow inside `base10` conversion
functions. The checks in place for overflow were not covering the
arithmetic operations correctly. Due to being a large number, this was
also causing a crash on ubsan, which was not producing a report.

This change also does an upstream change to make it possible to use
`base/numerics` for the necessary overflow checks over `int256_t`. A
CL[1] has been submitted upstream, to have this issue correct in
chromium. This is an incidental fix, and large numbers are not supported
by clang's builtins. Nonetheless this upstream change allows us to use
the slower path.

[1] https://crrev.com/c/5914985