Fix failure calling `json` function over `cxxbridge`
It is not entirely clear what is the underlying mechanism for this
failure, and it should be investigated further. On Windows Debug builds,
any calls to functions like `json::convert_all_numbers_to_string` were
causing a runtime crash. Bisecting our branch has revealed that this
crash started after an entirely unrelated change [1] that in fact seems
quite anodyne.

This change correct the runtime failure with a minimum change.

Resolves https://github.com/brave/brave-browser/issues/41497

[1] https://github.com/brave/brave-core/pull/25830