[ubsan] Avoid early downcast of `BraveTabStrip` (#25917)
`TabStrip` is both inherited by, and substituted with `BraveTabStrip`.
However, `TabStrip` instantiates `TabContainerImpl` and passes a pointer
to `this` the controller. `BraveTabContainer` is created at this point,
inheriting from `TabContainerImpl`, and during its constructor call, it
tries to pass the `TabStrip` pointer it has as `tab_slot_controller_`
as `BraveTabStrip` to some member overriden in the body of
`TabStripLayoutHelper` for future use.

In summary, when this downcast to `BraveTabStrip` occurs in
`BraveTabContainer`, the pointer has a vtable still for `TabStrip`, as
`BraveTabStrip` is not fully constructed yet, and no calls to its
members can be made.

This PR changes the member we insert in `TabStripLayoutHelper` to be a
`TabStrip` ptr. This may look worse, but it will allow sanitisers to
catch any early downcasts that could however occur, as those would be
undefined behaviour.