[Android] Rename SettingsLauncher to SettingsNavigation
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/79741f6bd5052dbfb30025c44f4b6851cb664e38

[Settings] Rename SettingsLauncher to SettingsNavigation

Now that SettingsLauncher provides a method to go back to the previous
page, it makes more sense to replace "launcher" in its name to something
more generic.

This patch renames SettingsLauncher and its friends to
SettingsNavigation. Also, methods in the interface are renamed to drop
mentions of "activity" since they don't necessarily start a new activity
in the single-activity mode.

Summary of changes:
- SettingsLauncher -> SettingsNavigation (w/ variants)
- launchSettingsActivity -> startSettings
- createSettingsActivityIntent -> createSettingsIntent
- finishCurrentFragment -> finishCurrentSettings

Bug: b/356743945