[Android] Password settings crash fix
Chromium changes:
https://source.chromium.org/chromium/chromium/src/+/df77cbacb9476b0b6d3e75cc34b5d127985e79c3

[Settings] Enable the single-activity mode by default

Bug: b/356743945

https://source.chromium.org/chromium/chromium/src/+/cd5fd3e6bcef39a904578857a4c1ab2419efacec

[Settings] Add a method to finish the current fragment

Some settings page need to finish when the user is done with them.
Previously we used getActivity().onBackPressed() to perform this, but
we might want to handle finish and back gesture differently, and also it
might have unintended interaction with metrics.

This patch introduces a canonical method for settings fragments to
declare that they are finished and the settings activity should show the
previous fragment.

We added the new method to SettingsLauncher because existing methods are
somewhat similar to the new one (handling navigation). In a later change
we plan to rename it to reflect the new addition.

Bug: b/356743945