Clear group id in-advance when tab is moved to pinned container
Our vertical tab uses CompoundTabContainer.
When tab is moved from the group by pinning, it's moved to
pinned TabContainerImpl before its tab group id is cleared.
And it causes runtime crash as using this tab from pinned TabContainerImpl
has assumption that it's not included in any group.
So, clear in-advance when tab enters to pinned TabContainerImpl.