(Android) Redesign private tab UI (#25850)
🎨 Redesign private tab UI following Figma design.

Private tab redesign includes three new sections "Leave no trace", "Browsing visibility" and "Brave VPN".
Brave VPN section will show a call to action text view that will redirect to the subscription page. If the user is already a paying customer, or the VPN is not supported the CTA view will be hidden.

The design is responsive and adjust to large device screens and tablets, following the logic provided by Figma design.
The entire layout is centered in the visible part of the page.