Add KDF implementations without CHECKs.
Chromium implementation of DeriveKeyScrypt has a CHECK in it that would
crash the browser if the underlying BoringSSL function failed. We use
this code from HDKey::GenerateFromV3UTC where the data is retrieved from
JSON and can't be guaranteed to always be valid. In order not to crash,
added DeriveKeyScrypt implementation without the CHECK and also changed
our DeriveKeyPbkdf2HmacSha256 implementation to no use CHECK either.

Additionally, moved HDKey::GenerateFromV3UTC code to use DeriveKeyScrypt
and DeriveKeyPbkdf2HmacSha256 directly instead of using SymmetricKey
wrappers that are being deprecated.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/4046edfc95050796d4a1bc54cdba6942257b6f38

commit 4046edfc95050796d4a1bc54cdba6942257b6f38
Author: Elly <ellyjones@chromium.org>
Date:   Thu Oct 10 20:51:36 2024 +0000

    crypto: add kdf interface

    Right now, //crypto supports KDFs via two interfaces:
    * PBKDF2-HMAC-SHA1 and Scrypt on crypto::SymmetricKey
    * HKDF-SHA256 via free functions

    This change:
    * Adds crypto::SubtlePassKey for access to subtle crypto primitives
    * Creates a new KDF interface
    * Moves the PBKDF2-HMAC-SHA1 and Scrypt implementations to that
      interface, and guards them with SubtlePassKeys
    * Adds new unit tests covering those KDFs
    * Deprecates SymmetricKey
    * Deletes many of the unit tests for SymmetricKey, which either covered
      logic which is now moved to the KDF implementations (and is covered by
      their tests) or tested key lengths we don't allow in production.

    Bug: 369653192, 369649346