Fix crash due to insufficient tokens
When calculating how much content the user's model can handle, we allocate a portion of the tokens to the prompt, and to the model's next response. At the time of this writing, that equals to 700 tokens. We cannot let the user set the context size lower than this allocation. As such, the kMinCustomModelContextSize needs to be based not on the number 1, but rather on the minimum number of tokens we expect to be used in a single interaction.