[CodeHealth] Use span of bytes to read arbitrary data (#25978)
This PR touches in different places of the code base where the use of
`reinterpret_cast` was necessary to allow data to be handled in
different format. Span is both safer and more ergonimic. It is
preferrable over `void*`, and particularly we should avoid any type of
casts involving byte values to numeric values, as those can cause
undefined behaviour.

This change additionally refactors `DeviceIdImpl::IsValidMacAddress` to
use `base::MakeFixedFlatSet` for the set of invalid mac addresses we
were searching for, as the search was sequential, and for successful
values it meant visiting every element in the array to check for
comparison. To make sure the binary search works as expected, a test has
been added.