Removed using SidePanelCoordinator::OnEntryWillDeregister
fix https://github.com/brave/brave-browser/issues/41748

In cr132, SidePanelCoordinator::OnEntryWillDeregister() is removed.
We relied on it to deregister observed entry from SidebarContainerView.
Instead, we'll use OnEntryWillHide() callback. It'll be called when
side panel is closed. At that time, we could remove observing it safely,
and will start to observe again if that entry is still live in tab's
registry.