Keeps MACL updated on model-changes
Now that context size can be set by the user, its enitrley possible that an update to the model options will include an updated context size, and as a result, an updated max_associated_content_length. As such, we should update the value here. We could drop this, and always pull the MACL from the model_options_ member, but I decided to keep this approach for parity with other engine consumers.