[iOS] Ensure bundled favicon initialization happens only once
If a number of bundled icons were requested all at once it was possible for it to be initialized multiple times which could lead to the `bundledIcons` state to be mutated multiple times across threads and cause a crash. This change ensures that we only ever initialize and set the `bundledIcons` state once.