[BYOM] Clear model index on save (#25984)
Clear model index on save

When we save the changes to a selected model, we need to not only hide the config UI, but we also need to set the active model index back to null. This prevents the next instance of the config UI from loading unrelated model details.