Add new purchase state OUT_OF_CREDENTIALS for VPN service
This handles the condition where the all the credentials for a given day
are redeemed and there are none left. This shouldn't happen - but can
happen if there are network issues. Specifically, when the client doesn't
receive response from server after redeeming the credential.

Fixes https://github.com/brave/brave-browser/issues/33031