Adds token-estimate for prompts
I thought it would be helpful to give the user an idea as to how many tokens will be used up by their system prompt. In this change, we assume a 4/1 (chars to tokens) relationship, and calculate the tokens-used estimate for the user's prompt. This estimate is shared next to the label/tooltip, above the textarea.