[BYOM] Custom System Prompts (#25727)
* Introduces support for custom system prompts

Introduces support for custom system prompts (resolving #40690). When set, the custom system prompt will replace Brave's default prompt at the start of each conversation.

There might be a bug with cr-textarea. Easiest solution right now is literally to move the textarea to another location (see https://github.com/brave/brave-core/pull/25727#issuecomment-2421332343).

I thought it would be helpful to give the user an idea as to how many tokens will be used up by their system prompt. In this change, we assume a 4/1 (chars to tokens) relationship, and calculate the tokens-used estimate for the user's prompt. This estimate is shared next to the label/tooltip, above the textarea.