[ubsan] Fix `SplitViewBrowserData` null deref
All browser tests for this type are incurring in a null deref, as
`browser` was not being passed into the type. This fix corrects these
cases.

Additionally, this change removes the unnecessary checks in
`SplitViewBrowserData` for things like `tab_strip_model`, and
`tab_strip_model_adapter_`, as these should never be nullable in this
class' logic.

https://github.com/brave/brave-browser/issues/41740