fixup! [Android] Fix for unused private tab resource
We don't use it since we have our own private tab UI.

Chromium change:
https://source.chromium.org/chromium/chromium/src/+/11bd4c71e1f22d72e63396f8b47ec3a5ab1957c7

Reland "Use Chrome's fading shadow implementation for IncognitoNTP"

This is a reland of commit 4db7d8944ecd4c49a744497b20a94d74687f3b6b

The CL was causing device offline on Android 13+ due to crash from
emulator GPU mode swiftshader. We have switched the GPU mode to
swangle on Android 13 and 15, and the CL runs fine on these builders.

(Android 14 builder is not switched. But it is not gardened yet so
OK to leave it as broken, and it is easy to make the switch)

Original change's description:
> Use Chrome's fading shadow implementation for IncognitoNTP
>
> Replace the view's fading shadow into Chrome's FadingShadowView and
> FadingEdgeScrollView. This is a workaround that #requiresFadingEdge
> doesn't work with #clipToPadding for ScrollView.
>
> Before (not edge to edge):
> https://screenshot.googleplex.com/9XESPLDvzviZ5cJ
> After:
> https://screenshot.googleplex.com/83Kd2unzVMBjEvr
>
> Bug: 365607068
> Change-Id: Ia156e10712de32f7141f379e1926fbcb3251caf0
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5846890
> Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
> Reviewed-by: Christian Dullweber <dullweber@chromium.org>
> Reviewed-by: Lijin Shen <lazzzis@google.com>
> Commit-Queue: Wenyu Fu <wenyufu@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1357331}

Bug: 365607068, 370084605