Internationalization Support for Leo (#25931)
Brave browser sends up a `selected_language` and a `system_language` with prompts.
The `selected_language` is initially an empty string, but the server can
send a SelectedLanguage event to set the language for future.
The server sets a selected_language based on what it detects the first
message's language to be in.  If it can't make that determine or that
message is not long enough it can fallback to using the system language.
Or if there is no system language specified it will use English.

More details here:
https://docs.google.com/document/d/1IWQNX-E40ujaQFQVavZEtgK3J068PgZ_xDbSLDWZOD8/edit?tab=t.0#heading=h.miolul91b7va