[CodeHealth] Use `SafeBaseName` for `TorConfig`
The use of `base::FilePath` in `TorConfig` could be a serious source for
a vulnerability if someone got control of the browser process. This
change employs the use of `SafeBaseName`, which only allows for the base
component of a path. With this browser process and the tor launcher do
agree over a certain definition of how the files are discovered.

Resolves https://github.com/brave/brave-browser/issues/41834