Move Brave strings to different number range
We previously stored Brave's strings after all of Chromium's resource strings.
This presents two problems, however:

1 - Every time new strings are added to Chrome in such a way that they conflict
with our strings, we need to increment our base.

2 - We are rapidly running out of room at the upper end of the range, as pak
files only support identifiers up to 65536.

In order to deal with this, I've moved Brave's strings to a section of the
Chromium identifiers that's reserved for ChromiumOS/Ash which we don't use.