[DanglingPtr][linux] Dangling violations as untriaged
This PR marks all the dangling pointers causing test crahses on linux as
`DanglingUntriaged`. This is the first step to gradually permit us to
switch on the our dangling pointer check in our CI. This change does not
include upstream patches, that may be required depending on how
difficult it becomes to fix those.

Resolves https://github.com/brave/brave-browser/issues/41896