[iOS] Fix playback speed + repeat mode buttons in Playlist fullscreen
There is unfortunately a SwiftUI regression in iOS 18 which causes `simultaneousGesture` to break the primary tap gesture on `Menu(_:label:primaryAction:)` views. This change swaps the fullscreen buttons that used to have a long-press gesture back to standard Button's that cycle but have no long-press ability to choose an explicit speed/repeat mode.