[DanglingPtr] Fix UaFs with `SidePanelEntry` observations
This has been lingering in the codebase for a while, due the reliance on
`MultisourceObservation` to keep track of the existing `SidePanelEntry`
observations in `SidebarContainerView`, and deregister them during
destruction. This runs into issues with the lifetime model currently
used by upstream, where observers of `SidePanelEntry` are only supposed
to add/remove themselves from observing, but at the same time be
prepared to forget about active observations, because there can be no
assumption about when the entry is going out of scope.

This change follows the chromium model, and merely uses
`AddObservation(this)` with no concern of whether or not the observation
could dangle.