[DanglingPtr] Fix UaFs with `SidePanelEntry` observation
This change addresses the uaf cases we were having with the multisource
observation for `SidePanelEntry` on `SidebarContainerView`. The main
issue we were having is that we can't precisely be notified of when a
given `SidePanelEntry` is going out of scope, and in many cases the
instances would be destroyed before the lifetime of
`SidebarContainerView`. In such cases, the observation would be kept
around in the multisource observer, and eventually would cause an UaF,
as the multisource observer was being destroyed and trying to remove the
outstanding observations that were dangling.

`SidePanelEntry` observations do dangle in upstream. That's the model
they have for the lifetimes: FeatureFoo is destroyed before
`SidePanelEntry`. There are exactly 2 ways for a `SidePanelEntry` to be
destroyed in chromium: FeatureFoo destroyes it via Deregister, or
`SidePanelRegistry` is destroyed before `FeatureFoo`.
`SidebarContainerView` cannot in a reliable way keep track of these
guarantees.

This change is predicated on the fact that `SidebarContainerView` will
always outlive all the use of events by any living `SidePanelEntry`
instance.

In this change we introduce a customisation point to `SidePanelEntry`
that allows us to check if a certain entry is being observed or not. If
it is we don't double-observe. Observations are then removed when a tab
is destroyed, which includes when it gets moved to another Browser
window.

Resolves brave/brave-browser#39053
Resolves brave/brave-browser#41924