[DanglingPtr] Enable dangling pointer detector for Linux (#26293)
This change switches on the dangling pointer detector for Linux. It also
includes a couple patches for chromium, that will have to be
investigated and corrected. With the detector on for Linux, there will
be a first line of bug prevention with this instrumentation, as it will
act as the first barrier for new cases of dangling pointers being
introduced in our codebase.

Resolves https://github.com/brave/brave-browser/issues/41955