[DanglingPtr][wallet] Prefer non-nullable states pt.2 (#26343)
This change touches several parts of wallets where we are passing
pointers around that are expected to be in a valid state, and remain so
for the duration of the lifetime of the object receiving them.

Historically, chromium had a guideline that required passing non-owning
references as raw pointers. This has changed for a while now, and there
are even supporting types like `raw_ref` for it. Raw pointers should be
used for cases where flexibility is required about the lifetime of what
is being pointed at. However, we should avoid what is in fact an
anti-pattern, of passing pointers for the lifetime of an object, and
then CHECKing the pointer for its validity.

This change also removes certain uses of `std::unique_ptr` that suffers
of the same issues.

Resolves https://github.com/brave/brave-browser/issues/42016