[Android] Moved back simplified view prompt for regular tabs
This is regression from this chromium change:
https://chromium.googlesource.com/chromium/src/+/c20ba57087f147b22a78a0fd916fdc4b887c2946

Cleanup Reader mode contextual page action

This CL cleans up ContextualPageActionReaderMode feature. There are
some details after the cleanup which can be summarized as :
1. Reader mode message UI is now only shown in CCT and incognito.
2. Finch param reader_mode_session_rate_limiting is set to
false. Hence it was removed in this CL.
3. CPA alternative color is also unused, but there are some feature
teams such as shopping who are interested in a similar experiment in
future. Hence it was kept as is.
4. There is also some code related to muting certain sites if the
user didn't take action. These are still needed for the message UI
that is shown for CCT and incognito.

Bug: 356237338