[cr132] `IsURLAllowedInIncognito` dropped `BrowserContext`
After discussion, we decided to eliminate the wallet check from
IsURLAllowedInIncognito instead allowing brave_wallet::IsAllowedForContext to
perform the required check. This does mean that if the "Allow wallet URLs in
private windows" option is not enabled and a user opens a wallet URL in a
private window it will result in a load failure rather than redirect to a normal
window.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/00708f413a9c2b37699a92cfadcdd18d872ffa2a

commit 00708f413a9c2b37699a92cfadcdd18d872ffa2a
Author: Thomas Lukaszewicz <tluk@chromium.org>
Date:   Mon Oct 21 19:41:59 2024 +0000

    Eliminate unused BrowserContext* param from IsURLAllowedInIncognito()

    Removes the param from the function and cleans up client code.

    Bug: 368059979