Simplify `gn` handlebars patching (#26365)
* [CodeHealth][rust] Simplify gn handlebars patching

This aligns our own handlebars file closer to the upstream one, so it is
easier to compare changes, and therefore to fix recent additions.

* [CodeHealth][rust] Brave `BUILD.gn` handlebars changes

This change contains the handlebars changes we apply to our own
handlebars file. This commit also adds a patch that can be used to
update the handlebar files, based on new upstream additions.