[CodeHealth][rust] `gnrt` run with the latest handlebars
During a recent PR to fix the rust handlebars for gn files, there was a
change to the copyright notice, and then this got overlooked, with the
new handlebars used to regenerate all `BUILD.gn` files with the new
notice. This has rendered a dirty `gnrt` run. This PR adds the updated
files.

This PR also moves the file to its correct place.

[1]https://github.com/brave/brave-core/pull/26365