[CodeHealth] Fix dangling span references
This PR fixes cases where a local constant is being used to initialise a
`span` but then the span is being used by the time the local has already
been discarded. This is undefined behaviour although in both cases it
seems to be incidently working.

Resolve https://github.com/brave/brave-browser/issues/42185