[CodeHealth] Drop `proc-macro-error` from `filecoin-cxx` (#26427)
* [CodeHealth] Drop `proc-macro-error` from `filecoin-cxx`

This PR updates all crates around `multihash` to use the latest
`multihash` version that does not depend anymore on `proc-macro-error`.
This crate is being removed as it is getting flagged in security audits
for being unmaintained, and it also relies on unstable rustc features.

* [rust] `gnrt` crate changes for `multihash` v.0.19

This commit includes all the mechanical changes done by `gnrt`.