[DanglingPtr] Remove `ProfileMenuViewBase` patching (#26499)
This patch was introduced as a temporary measure to allow switching on
the dangling pointer detector.

This PR fixes the two dangling pointers in `ProfileMenuViewBase` by
avoiding calling `SetProfileIdentityInfo` more than once, and actually
overriding it, so the first call to the `SetProfileIdentityInfo` can be
given the correct arguments.

One test had to be disabled, as it was not actually testing the way the
profile menu viewer was being instantiated. The way brave does it
requires to consult theme data, which is not available for this
particular test.

Resolves https://github.com/brave/brave-browser/issues/42231