[CodeHealth] Fix various use-after-move violations (#26526)
This change fixes several instances where use-after-move is possible in
our codebase. For potential use-after-move cases involving
`OnceCallback` instances, those are potential crahes in the application.

Additionally, this PR corrects the use of moved-from std containers. It
is valid to use a moved-from container, but it is expected that the
container is cleared before being used again.

Resolves https://github.com/brave/brave-browser/issues/42264