[CodeHealth] Use `//base` backend for number conversions
Recently we managed to get a fix upstream to allow safe numerics with
`int256_t` and similar types. This was done to correct some undefined
behaviour on implemention doing back and forth convertion between
256-bit numbers and string. Nevertheless, as safe numerics can now be
used with these large ints, it also allows us to just use the `//base`
backend for things like base10 conversions.

This PR removes our own implementations for these conversions, and moves
to use the chromium ones.

Resolves https://github.com/brave/brave-browser/issues/42300