Fix likely dangling span use (#26583)
This change loads the value return by `Decrypt*` into a local variable
so it can use the span inside `EXPECT_EQ`. It is not guaranted that
`EXPECT_EQ` holds an expression lifetime, as the macro is expanded, so
the storoge of vector could be gone by the time the span is being read.