Send UpdaterName in Component Updater Requests
The update protocol asks client implementations to self-identify so that issues with one client implementation aren't mistaken for issues with another.

Brave reports an empty string as its identifier, which is not an ideal choice. (It's uncommon enough to not provide practical privacy benefits, especially when combined with Brave's versioning scheme, but obfuscated enough to add hours of friction to any escalations from server maintainers to Brave maintainers.)

This commit changes Brave's component updater to name itself "BraveComponentUpdater" when communicating to update servers.