[DanglingPtr] Remove patching for `AutocompleteMatch`
This PR removes the patching for `AutocompleteMatch`, introduced to
allow switching on the dangling pointer detector, as a temporary
measure.

The main issue was that `BraveRenderViewContextMenu` was instantiating
`AutocompleteClassifier` discretely, rather than using the factory for
it, and this was causing dangling checks to fire with
`AutocompleteClassifier`, which is supposed to be retrieved with
`AutocompleteClassifierFactory`.

Resolves https://github.com/brave/brave-browser/issues/42339