[CodeHealth] Avoid constant arrays of string pointers (#26582)
This PR changes several places where the construct constant arrays for
string pointers are being (`constexpr const char* kValue ...`). This
makes code overall safer. In most cases, a `std::array` of `string_view`
is fine. However, in some places a fixed flat set is the better choice,
as we are doing searches in some of these arrays.