[Android] New param for TabbedRootUiCoordinator c-tor
Chromium change:
https://chromium.googlesource.com/chromium/src/+/272a97cf5dd0cb0319ce98fd6c726e4c47beda76

Create EdgeToEdgeSystemBarColorHelper in the E2EManager

With this change, the E2EManager will create the EdgeToEdgeSystemBarColorHelper for managing system bar coloring edge-to-edge. EdgeToEdgeSystemBarColorHelper will receive a supplier from ChromeBaseAppCompatActivity through the E2EManager for coloring Chrome-based system bars (as opposed to the OS system bars from the Window) - for most activities, this will be the EdgeToEdgeLayout if the edge-to-edge-everywhere flag is enabled, though the ChromeTabbedActivity will instead supply the bottom chin.

Currently, this doesn't fully support the status bar color, the EdgeToEdgeSystemBarColorHelper will have to be expanded to fully account for the status bar and top insets.

Bug: 377959835