Fixed dangling TabInterface from shared pinned tab feature
Whenever tab is about to be deleted, proper method should be called
from TabStripModel for deleting. And TabModel::WillDetach() should
be called with kDelete reasion. So far, we just call DetachTabAtForInsertion()
and did nothing more when browser window is closing when shared pinned tab
feature is enabled. Because of that, EmbeddingTabTracker doesn't get
chance to cleanup TabInterface pointer.

Also when tab is discared, old tab's TabInterface should be cleared.