[CodeHealht][wallet] Avoid `NoDestructor` in `eth_data_builder.cc`
The use of `NoDestructor` for a map of strings in `ChainIdToVersion` is
wasteful of resources, when the actual data can be seen at compile time.

This PR converts this static `std::map` into a fixed flat map. This
required changing the type of the elements to `string_view`, so this PR
goes a bit the extra mile to modernise some of the functions around to
use `string_view`.

Resolve https://github.com/brave/brave-browser/issues/42432