[CodeHealth][wallet] Remove `NoDestructor` in `GetRecordKeys`
This function has been returning a vector of strings, to be used by
`GetMany`, and therefore a `NoDestructor` instance is needed. However,
the only reason why this has to be a vector of strings is because in the
end `EncodeStringArray` requires a vector of strings.

This PR changes `EncodeStringArray` to take spans of strings and of
string views. This will allows to convert the keys into a `constexpr`
value, and avoid the use of `NoDestructor`.

In a follow up PR, additional callsites of `EncodeStringArray` may be
migrated to the `string_view` variant.

This is a follow up of these PRs:
https://github.com/brave/brave-core/pull/26683
https://github.com/brave/brave-core/pull/26688

Resolves https://github.com/brave/brave-browser/issues/42453