[spanify] Use safe buffers pt.3
This change avoids certain uses of `std::next`/`prev`, or in certain
cases simplifies their use, to make it more obvious what certain ranged
operations are actually doing.

Most changes here are standard. However, with `ReorderMatch` in
particular, a rewrite was done, as `ReorderMatch` was to obscure, and it
was doing two distinct things at the same time.

Resolves https://github.com/brave/brave-browser/issues/42489