[CodeHealth] Normalise `const` declaration pt.1
This PR is part of a couple of rewrites to be carried out normalising
the cases where we have constant declaration. In this PR we have the
following fixes:

 - Remove redundant `const` qualifier alongside `constexpr`
 - Function-scope constants as `static constexpr`
 - Remove the use of `extern` declarations
 - Whenever possible use `constexpr`
 - When appropriate, use `MakeFixedFlatSet`.

These changes are in conformance with some upcoming clarified guidelines.
Furthermore, this takes us one step closer to have some automation for
clang-tidy, as right now there are too many clang-tidy warnings around
constant/variable naming.

Resolves https://github.com/brave/brave-browser/issues/42490