[ubsan] Invalid cast fixes for upstream tests
This change delays casting in a few places, as this is now breaking with
certain upstream tests that seem to inject different types in the cast
pointers. This in itself is not a problem provided that those upstream
tests are not running through brave code that relies on those cast.

Resolves https://github.com/brave/brave-browser/issues/42491