[iOS] Fix using embedded DarkReader for night mode
This change fixes night mode by using the `DarkReader.setFetchMethod` to ensure CORS requests load correctly. It also removes the list of legacy night mode blocked urls as it's no longer needed anyways and was causing a bug that any URL with a `x` in the host or path would be run dark mode on it.