[iOS] Remove explicit device selection in Fastlane definition & iPad run
This change removes the iPad test plan & removes the explicit device selection from the Fastlane configuration so that Fastlane can automatically pick the simulator to run it on based on the available & compatible simulators. This also removes the iPad test plan and adjusts the UserAgent code & unit tests to test against the pref + idiom directly for iPad.